文字数カウントを別ファイルに移動した
